app/import/12_html_generator/20250515/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
anthropic>=0.51,<1
gunicorn
pytest==8.1.1
pytest-cov==5.0.0
